moved conditions for enhanced funding to the enhanced funding folder
